+++ collector/src/main/java/com/bitmovin/analytics/ads/AdPosition.kt
@@ -1,8 +1,12 @@
 package com.bitmovin.analytics.ads
 
 @Suppress("ktlint")
-enum class AdPosition(name: String) {
-    pre("pre"),
-    mid("mid"),
-    post("post")
+enum class AdPosition(val position: String) {
+    PRE("pre"),
+    MID("mid"),
+    POST("post");
+
+    override fun toString(): String {
+        return position
+    }
 }