+++ collector/src/main/java/com/bitmovin/analytics/ads/AdTagType.kt
@@ -1,8 +1,13 @@
 package com.bitmovin.analytics.ads
 
-enum class AdTagType(name: String) {
+@Suppress("ktlint")
+enum class AdTagType(val type: String) {
     VAST("vast"),
     VMAP("vmap"),
     VPAID("vpaid"),
-    UNKNOWN("unknown"),
+    UNKNOWN("unknown");
+
+    override fun toString(): String {
+        return type
+    }
 }